Remove dedicated orders response object. Remove old test files. Add date test.

+++ src/main/java/com/musala/task/dao/orders/TomatoOrderDAO.java
@@ -5,5 +5,11 @@ import java.util.List;
 import com.musala.task.model.TomatoOrder;
 
 public interface TomatoOrderDAO {
-	public List<TomatoOrder> getTomatoOrders(long size);
+	
+	/**
+	 * Get orders from db.
+	 * @param size - number of orders to take from db.
+	 * @return
+	 */
+	public List<TomatoOrder> getTomatoOrders(int size);
 }
