+++ src/main/java/com/musala/task/services/orders/TomatoOrdersService.java
@@ -5,5 +5,10 @@ import java.util.List;
 import com.musala.task.model.TomatoOrder;
 
 public interface TomatoOrdersService {
-	public List<TomatoOrder> retrieveTomatoOrders(long size);
+	/**
+	 * Retrieve orders from db.
+	 * @param size - the number of orders to retrieve.
+	 * @return
+	 */
+	public List<TomatoOrder> retrieveTomatoOrders(int size);
 }
